setup structure for developing the pages
