+++ services/web/src/website/HackathonPage/HackathonPage.tsx
@@ -1,0 +1,3 @@
+export const HackathonPage = () => {
+    return <h1>HackPage</h1>;
+};
