+++ services/web/src/website/MainPage/MainPage.tsx
@@ -1,0 +1,3 @@
+export const MainPage = () => {
+    return <h1>MainPage</h1>;
+};
